+++ src/containers/internal/views/recent-calls/jaeger/bar.tsx
@@ -2,27 +2,11 @@ import React from "react";
 import { JaegerGroup } from "src/api/jaeger-types";
 
 import "./styles.scss";
+import { formattedDuration } from "./utils";
 
 type BarProps = {
   group: JaegerGroup;
   handleRowSelect: (grp: JaegerGroup) => void;
-};
-
-const formattedDuration = (duration: number) => {
-  if (duration < 1) {
-    return (Math.round(duration * 100) / 100).toFixed(2) + "ms";
-  } else if (duration < 1000) {
-    return (Math.round(duration * 100) / 100).toFixed(0) + "ms";
-  } else if (duration >= 1000) {
-    const min = Math.floor((duration / 1000 / 60) << 0);
-    if (min == 0) {
-      const secs = parseFloat(`${duration / 1000}`).toFixed(2);
-      return `${secs}s`;
-    } else {
-      const sec = Math.floor((duration / 1000) % 60);
-      return `${min}m ${sec}s`;
-    }
-  }
 };
 
 export const Bar = ({ group, handleRowSelect }: BarProps) => {
